chore(server): :alien: update airtable api auth
