+++ server/cloudbuild.yaml
@@ -34,7 +34,7 @@ steps:
         "--memory",
         "1G",
         "--set-env-vars",
-        "API_KEY=sm://refugies-info/API_KEY,API_SECRET=sm://refugies-info/API_SECRET,GCLOUD_CLIENT_EMAIL=sm://refugies-info/GCLOUD_CLIENT_EMAIL,GCLOUD_CLIENT_ID=sm://refugies-info/GCLOUD_CLIENT_ID,GCLOUD_CLIENT_X509=sm://refugies-info/GCLOUD_CLIENT_X509,GCLOUD_PKEY=sm://refugies-info/GCLOUD_PKEY,GCLOUD_PRIVATE_KEY_ID=sm://refugies-info/GCLOUD_PRIVATE_KEY_ID,MONGODB_URI=sm://refugies-info/${_MONGODB_VAR},NODE_ENV=${_NODE_ENV},OVH_PASS=sm://refugies-info/OVH_PASS,REACT_APP_SITE_SECRET=sm://refugies-info/REACT_APP_SITE_SECRET,SECRET=sm://refugies-info/SECRET,TTS_KEY_1=sm://refugies-info/TTS_KEY_1,accountSid=sm://refugies-info/accountSid,airtableApiKey=sm://refugies-info/airtableApiKey,authToken=sm://refugies-info/authToken,CLOUD_NAME=sm://refugies-info/CLOUD_NAME,AIRTABLE_BASE_USERS=sm://refugies-info/AIRTABLE_BASE_USERS,SENDGRID_API_KEY=sm://refugies-info/SENDGRID_API_KEY,POSTMAN_SECRET=sm://refugies-info/POSTMAN_SECRET,CRON_TOKEN=sm://refugies-info/CRON_TOKEN,AIRTABLE_API_KEY_APPLI=sm://refugies-info/AIRTABLE_API_KEY_APPLI,AIRTABLE_BASE_DIAIR_TEST=sm://refugies-info/AIRTABLE_BASE_DIAIR_TEST,AIRTABLE_BASE_TRAD=sm://refugies-info/AIRTABLE_BASE_TRAD,FRONT_SITE_URL=sm://refugies-info/${_FRONT_URL_VAR},ALGOLIA_API_KEY=sm://refugies-info/ALGOLIA_API_KEY,ALGOLIA_INDEX=sm://refugies-info/${_ALGOLIA_INDEX_VAR},EXPO_ACCESS_TOKEN=sm://refugies-info/EXPO_ACCESS_TOKEN,SENDINBLUE_API_KEY=sm://refugies-info/SENDINBLUE_API_KEY,SLACK_WEBHOOK_URL=sm://refugies-info/SLACK_WEBHOOK_URL",
+        "API_KEY=sm://refugies-info/API_KEY,API_SECRET=sm://refugies-info/API_SECRET,GCLOUD_CLIENT_EMAIL=sm://refugies-info/GCLOUD_CLIENT_EMAIL,GCLOUD_CLIENT_ID=sm://refugies-info/GCLOUD_CLIENT_ID,GCLOUD_CLIENT_X509=sm://refugies-info/GCLOUD_CLIENT_X509,GCLOUD_PKEY=sm://refugies-info/GCLOUD_PKEY,GCLOUD_PRIVATE_KEY_ID=sm://refugies-info/GCLOUD_PRIVATE_KEY_ID,MONGODB_URI=sm://refugies-info/${_MONGODB_VAR},NODE_ENV=${_NODE_ENV},OVH_PASS=sm://refugies-info/OVH_PASS,REACT_APP_SITE_SECRET=sm://refugies-info/REACT_APP_SITE_SECRET,SECRET=sm://refugies-info/SECRET,TTS_KEY_1=sm://refugies-info/TTS_KEY_1,accountSid=sm://refugies-info/accountSid,authToken=sm://refugies-info/authToken,CLOUD_NAME=sm://refugies-info/CLOUD_NAME,AIRTABLE_BASE_USERS=sm://refugies-info/AIRTABLE_BASE_USERS,SENDGRID_API_KEY=sm://refugies-info/SENDGRID_API_KEY,POSTMAN_SECRET=sm://refugies-info/POSTMAN_SECRET,CRON_TOKEN=sm://refugies-info/CRON_TOKEN,AIRTABLE_TOKEN=sm://refugies-info/AIRTABLE_TOKEN,AIRTABLE_BASE_DIAIR_TEST=sm://refugies-info/AIRTABLE_BASE_DIAIR_TEST,AIRTABLE_BASE_TRAD=sm://refugies-info/AIRTABLE_BASE_TRAD,FRONT_SITE_URL=sm://refugies-info/${_FRONT_URL_VAR},ALGOLIA_API_KEY=sm://refugies-info/ALGOLIA_API_KEY,ALGOLIA_INDEX=sm://refugies-info/${_ALGOLIA_INDEX_VAR},EXPO_ACCESS_TOKEN=sm://refugies-info/EXPO_ACCESS_TOKEN,SENDINBLUE_API_KEY=sm://refugies-info/SENDINBLUE_API_KEY,SLACK_WEBHOOK_URL=sm://refugies-info/SLACK_WEBHOOK_URL",
         "--set-env-vars",
         "MINIMUM_APP_VERSION=${_MINIMUM_APP_VERSION}",
         "--service-account",
