+++ server/example-env-file.env
@@ -1,10 +1,8 @@
 accountSid=demo
-airtableApiKey=demo
-AIRTABLE_BASE_DIAIR=demo
+AIRTABLE_TOKEN=demo
+AIRTABLE_BASE_DIAIR_TEST=demo
 AIRTABLE_BASE_USERS=demo
 AIRTABLE_BASE_TRAD=demo
-AIRTABLE_BASE_APPLI=demo
-AIRTABLE_API_KEY_APPLI=demo
 API_KEY=demo
 API_SECRET=demo
 authToken=demo
