--- server/src/workflows/needs/extractNeedsFichesFromAirtable/data.ts
@@ -1,277 +1,0 @@
-export const needsAirtable = [
-  {
-    needName: "Obtenir des aides financières (RSA, CAF)",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "rec6GkO7k9fa9l5ss",
-  },
-  {
-    needName: "Trouver un premier travail",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "recIlK8bAVaoMOLsv",
-  },
-  {
-    needName: "Être accompagné pour trouver un travail",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "rec9KBmGAp3Jt5mPl",
-  },
-  {
-    needName: "Où chercher un travail ?",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "recXWhLi2JKS3yRiC",
-  },
-  {
-    needName: "Savoir candidater (CV, entretien...)",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "reclUMfvv31GO1fNY",
-  },
-  {
-    needName: "Commencer à apprendre le français",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "recbUKtrnqx80IOU7",
-  },
-  {
-    needName: "Apprendre le français du monde du travail",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "rec04lZ7zj6beWg47",
-  },
-  {
-    needName: "Faire une activité culturelle ou créative",
-    tagRecordId: "recAUhODnxjDVTzrH",
-    needId: "rec6MgbmL7pCu0Roq",
-  },
-  {
-    needName: "Faire du sport",
-    tagRecordId: "recAUhODnxjDVTzrH",
-    needId: "recmWPo2wNNCiMsX6",
-  },
-  {
-    needName: "Découvrir ma ville",
-    tagRecordId: "recAUhODnxjDVTzrH",
-    needId: "recXw9ZQB1ozuMCqO",
-  },
-  {
-    needName: "Apprendre quand j'ai du temps libre",
-    tagRecordId: "recAUhODnxjDVTzrH",
-    needId: "recJe6TNSD6DiXNTg",
-  },
-  {
-    needName: "Tester mon niveau de langue",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "rec60F8gHDQgt5s1d",
-  },
-  {
-    needName: "Passer un diplôme (DILF, DELF, DALF, TCF)",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "recOKJbYvfQE2O8YM",
-  },
-  {
-    needName: "Approfondir mes compétences linguistiques",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "rec5EADtOnUCqguM8",
-  },
-  {
-    needName: "Préparer mon entrée à l'université",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "recWBk0ElGTJKzIvB",
-  },
-  {
-    needName: "Trouver les bonnes études",
-    tagRecordId: "recjfO6nHolodnwCU",
-    needId: "recyyzjXtOBYtomYL",
-  },
-  {
-    needName: "Être aidé pour réviser et apprendre",
-    tagRecordId: "recjfO6nHolodnwCU",
-    needId: "recKrdB9l4XhZBja3",
-  },
-  {
-    needName: "Avoir des tarifs réduits pour la culture",
-    tagRecordId: "recD80epQNMxd1WSH",
-    needId: "recc0uRCefDLa37ko",
-  },
-  {
-    needName: "Avoir le permis de conduire",
-    tagRecordId: "recYUccgpUBYsQVri",
-    needId: "recKViJClQmi5rBk3",
-  },
-  {
-    needName: "Avoir des tarifs réduits pour les transports",
-    tagRecordId: "recYUccgpUBYsQVri",
-    needId: "reciMfouglDduvL6i",
-  },
-  {
-    needName: "Être aidé pour faire mes démarches administratives",
-    tagRecordId: "recKCNdX5nozbZnfG",
-    needId: "reczjT6Kd6raDa3pl",
-  },
-  {
-    needName: "Obtenir mes papiers importants",
-    tagRecordId: "recKCNdX5nozbZnfG",
-    needId: "recvZYoxl8u2s3FcJ",
-  },
-  {
-    needName: "Les démarches OFPRA (état civil)",
-    tagRecordId: "recKCNdX5nozbZnfG",
-    needId: "recbTOYIVe6XIOkTJ",
-  },
-  {
-    needName: "Gérer mon budget et obtenir des aides",
-    tagRecordId: "recKCNdX5nozbZnfG",
-    needId: "recFd0bpa97vQzfpu",
-  },
-  {
-    needName: "Les démarches pour ma famille",
-    tagRecordId: "recKCNdX5nozbZnfG",
-    needId: "recXio2uFMUWgoiP8",
-  },
-  {
-    needName: "Ouvrir un compte bancaire",
-    tagRecordId: "recKCNdX5nozbZnfG",
-    needId: "recQdTmCKszGGI3W3",
-  },
-  {
-    needName: "Obtenir mes papiers importants de santé",
-    tagRecordId: "recR2xYxNUTeQ0mlW",
-    needId: "recQWvBiwcKoBXaJy",
-  },
-  {
-    needName: "Me faire soigner gratuitement",
-    tagRecordId: "recR2xYxNUTeQ0mlW",
-    needId: "recqHYSQQhgTdUm2o",
-  },
-  {
-    needName: "Avoir une aide psychologique",
-    tagRecordId: "recR2xYxNUTeQ0mlW",
-    needId: "rec2IMApQrS5A0eo4",
-  },
-  {
-    needName: "Découvrir la culture en apprenant le français",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "reczCB0LPQVONRw9E",
-  },
-  {
-    needName: "Découvrir la culture sur internet",
-    tagRecordId: "recD80epQNMxd1WSH",
-    needId: "recC67DKMjBqSUCWs",
-  },
-  {
-    needName: "Trouver un logement",
-    tagRecordId: "recPtImEYKM3BR3OR",
-    needId: "recpvGZVq2P1ZOUgK",
-  },
-  {
-    needName: "Être accompagné pour trouver un logement",
-    tagRecordId: "recPtImEYKM3BR3OR",
-    needId: "rec6XtFpalZ32Qu5N",
-  },
-  {
-    needName: "Trouver un hébergement (courte durée)",
-    tagRecordId: "recPtImEYKM3BR3OR",
-    needId: "recxMgtuwMWgvgXEC",
-  },
-  {
-    needName: "Trouver un lieu de repos",
-    tagRecordId: "recPtImEYKM3BR3OR",
-    needId: "recWk2IIs0Fd4YcSH",
-  },
-  {
-    needName: "Être bénévole pour une association",
-    tagRecordId: "recHqPZrDSCuyi2hE",
-    needId: "recxGDQr5kgHUzdQG",
-  },
-  {
-    needName: "Faire un service civique",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "recb7OntylAbjZZpS",
-  },
-  {
-    needName: "Être accompagné pour trouver une formation",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recabFmsQ0vequhqc",
-  },
-  {
-    needName: "Faire reconnaître mon diplôme / VAE",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recNIt1CWDxsXclsi",
-  },
-  {
-    needName: "Apprendre à utiliser un ordinateur",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recRIU6C9AEC9nH6a",
-  },
-  {
-    needName: "Me former aux métiers du numérique",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "rec1VFZ1LGsCnLMGe",
-  },
-  {
-    needName: "Me former aux métiers de la cuisine",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recifejKWusD5MtqW",
-  },
-  {
-    needName: "Me former aux métiers du bâtiment (BTP)",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recEDhSZivHnlCz5M",
-  },
-  {
-    needName: "Me former aux métiers du social",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "rec8RUExqr2yWBlqx",
-  },
-  {
-    needName: "Me former aux métiers du sport",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recjMYlzYtbVwYOXf",
-  },
-  {
-    needName: "Me former aux métiers de l'agriculture",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recnqJ0wrfVoeBNd8",
-  },
-  {
-    needName: "Me former au métier de la surveillance",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "rec8La93gWOOjDRzl",
-  },
-  {
-    needName: "Financer ma formation",
-    tagRecordId: "recNRZArFePyPhuiJ",
-    needId: "recNto0HF4jVEk4MG",
-  },
-  {
-    needName: "Créer une entreprise",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "rec7OscWhT84JZjQX",
-  },
-  {
-    needName: "Avoir un ordinateur ou un téléphone pas cher",
-    tagRecordId: "recHUq0rw4WsBY3x2",
-    needId: "recQcnUlGsx6xoQjX",
-  },
-  {
-    needName: "M'entraîner à parler le français",
-    tagRecordId: "recewcmBlTHh2suZf",
-    needId: "recVyVvgSK9xs3pkU",
-  },
-  {
-    needName: "Me nourrir à petits prix",
-    tagRecordId: "recR2xYxNUTeQ0mlW",
-    needId: "recyNkGgdgBkFijSc",
-  },
-  {
-    needName: "Prendre soin de mon corps",
-    tagRecordId: "recR2xYxNUTeQ0mlW",
-    needId: "recfJSVbCDhBnvDAv",
-  },
-  {
-    needName: "Rencontrer de nouvelles personnes",
-    tagRecordId: "recXuT0VkRSRnEgQv",
-    needId: "recmZ1oJ2ptaY5Coe",
-  },
-  {
-    needName: "Voyager quand on est réfugié",
-    tagRecordId: "recYUccgpUBYsQVri",
-    needId: "rec2akAmeaZfYvMVA",
-  },
-];
