--- server/src/workflows/needs/extractNeedsFichesFromAirtable/index.ts
@@ -1,1 +1,0 @@
-export * from "./extractNeedsFichesFromAirtable";
