--- server/src/workflows/needs/retrieveNeedsFromAirtable/index.ts
@@ -1,1 +1,0 @@
-export * from "./retrieveNeedsFromAirtable";
